mpi/errhan: split MPIR implementations from MPI functions
Note: we also moved the prototype declarations from mpir_comm.h to
src/include/mpir_impl.h so we can better group them.
